fixed idle conn for multi db
